Introduces separate speech package.
