+++ src/main/java/com/progressiveaccess/cmlspeech/speech/EnglishAtomTable.java
@@ -25,7 +25,7 @@
 
 //
 
-package com.progressiveaccess.cmlspeech.sre;
+package com.progressiveaccess.cmlspeech.speech;
 
 
 /**
