+++ src/main/java/com/progressiveaccess/cmlspeech/speech/ExpertSpeechVisitorFactory.java
@@ -13,49 +13,49 @@
 // limitations under the License.
 
 /**
- * @file   AtomTableFactory.java
+ * @file   ExpertSpeechVisitorFactory.java
  * @author Volker Sorge <sorge@zorkstomp>
  * @date   Thu Jul 30 05:33:44 2015
  * 
- * @brief  Factory for generating atom tables.
+ * @brief  Factory for generating atom visitors.
  * 
  * 
  */
 
 //
-package com.progressiveaccess.cmlspeech.sre;
+package com.progressiveaccess.cmlspeech.speech;
 
 import java.util.HashMap;
 import java.util.Map;
 
 
 /**
- * Factory for generating language specific atom tables.
+ * Factory for generating language specific atom visitors.
  */
 
-public class AtomTableFactory {
+public class ExpertSpeechVisitorFactory {
 
-  private static Map<String, AtomTable> ATOM_TABLES;
+  private static Map<String, SpeechVisitor> VISITORS;
 
   static {
-    ATOM_TABLES = new HashMap<String, AtomTable>();
-    ATOM_TABLES.put("english", new EnglishAtomTable());
+    VISITORS = new HashMap<String, SpeechVisitor>();
+    VISITORS.put("english", new EnglishExpertSpeechVisitor());
   }
   
-  public static AtomTable getAtomTable(String language) {
-    AtomTable table = ATOM_TABLES.get(language);
-    if (table != null) {
-      return table;
+  public static SpeechVisitor getSpeechVisitor(String language) {
+    SpeechVisitor visitor = VISITORS.get(language);
+    if (visitor != null) {
+      return visitor;
     }
     switch (language) {
     case "japanese":
-      table = new JapaneseAtomTable(); 
+      visitor = new JapaneseExpertSpeechVisitor(); 
       break;
     default:
-      return ATOM_TABLES.get("english");
+      return VISITORS.get("english");
     }
-    ATOM_TABLES.put(language, table);
-    return table;
+    VISITORS.put(language, visitor);
+    return visitor;
   }
   
 }
