+++ src/main/java/com/progressiveaccess/cmlspeech/speech/JapaneseAtomTable.java
@@ -25,7 +25,7 @@
 
 //
 
-package com.progressiveaccess.cmlspeech.sre;
+package com.progressiveaccess.cmlspeech.speech;
 
 
 /**
