+++ src/main/java/com/progressiveaccess/cmlspeech/speech/JapaneseSimpleSpeechVisitor.java
@@ -13,19 +13,19 @@
 // limitations under the License.
 
 /**
- * @file   EnglishExpertSpeechVisitor.java
+ * @file   JapaneseSimpleSpeechVisitor.java
  * @author Volker Sorge
  *          <a href="mailto:V.Sorge@progressiveaccess.com">Volker Sorge</a>
- * @date   Tue Jun 30 14:46:54 2015
+ * @date   Wed Jul  8 09:07:00 2015
  *
- * @brief  Simple speech visitor.
+ *a @brief  Visitor for simple speech descriptions.
  *
  *
  */
 
 //
 
-package com.progressiveaccess.cmlspeech.sre;
+package com.progressiveaccess.cmlspeech.speech;
 
 import com.progressiveaccess.cmlspeech.analysis.RichStructureHelper;
 import com.progressiveaccess.cmlspeech.connection.Bridge;
@@ -38,6 +38,7 @@ import com.progressiveaccess.cmlspeech.structure.RichAliphaticChain;
 import com.progressiveaccess.cmlspeech.structure.RichAtom;
 import com.progressiveaccess.cmlspeech.structure.RichAtomSet;
 import com.progressiveaccess.cmlspeech.structure.RichBond;
+import com.progressiveaccess.cmlspeech.structure.RichChemObject;
 import com.progressiveaccess.cmlspeech.structure.RichFunctionalGroup;
 import com.progressiveaccess.cmlspeech.structure.RichFusedRing;
 import com.progressiveaccess.cmlspeech.structure.RichIsolatedRing;
@@ -45,21 +46,37 @@ import com.progressiveaccess.cmlspeech.structure.RichMolecule;
 import com.progressiveaccess.cmlspeech.structure.RichSetType;
 import com.progressiveaccess.cmlspeech.structure.RichSubRing;
 
+import com.google.common.base.Joiner;
+
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.Map;
 import java.util.SortedSet;
 import java.util.TreeSet;
 
 /**
- * Produces the basic speech for structures.
+ * Produces the simple speech for structures.
  */
 
-public class EnglishExpertSpeechVisitor extends AbstractSpeechVisitor {
+public class JapaneseSimpleSpeechVisitor extends AbstractSpeechVisitor {
 
   private boolean shortDescription = false;
+  private boolean subject = true;
+  private static Map<String, String> bondMap =
+      new HashMap<String, String>();
+
+  static {
+    bondMap.put("single", "単");
+    bondMap.put("double", "二重");
+    bondMap.put("triple", "三重");
+    bondMap.put("quadruple", "四重");
+  }
 
 
   @Override
   public void visit(final RichBond bond) {
-    this.addSpeech(bond.getName());
+    this.addSpeech(bondMap.get(bond.orderDescription()));
+    this.addSpeech("結合"); // bond
   }
 
 
@@ -71,8 +88,11 @@ public class EnglishExpertSpeechVisitor extends AbstractSpeechVisitor {
       this.describeSuperSystem(atom);
       return;
     }
-    this.addSpeech(atom.getName());
+    this.addSpeech(atom.getName()); // Done below
     this.addSpeech(position);
+    if (this.subject) {
+      this.addSpeech("は、"); // Separator (only after subject).
+    }
     if (this.shortDescription) {
       return;
     }
@@ -82,109 +102,150 @@ public class EnglishExpertSpeechVisitor extends AbstractSpeechVisitor {
 
   @Override
   public void visit(final RichIsolatedRing ring) {
-    this.addName(ring);
-    this.addSpeech("ring");
+    this.addSpeech(ring.getComponentsPositions().size());
+    this.addSpeech("員"); // Elements
+    this.addSpeech("環"); // Ring
+    this.addSpeech("、"); // Punctuation
     if (this.shortDescription) {
       return;
     }
+    this.describeReplacements(ring);
+    this.describeMultiBonds(ring);
     this.describeSubstitutions(ring);
   }
 
 
   @Override
   public void visit(final RichFusedRing ring) {
-    this.addSpeech("Fused ring system");
-    this.addName(ring);
+    this.addSpeech("縮合環系");  // Fused ring system
+    this.addSpeech("、"); // Punctuation
     if (this.shortDescription) {
       return;
     }
-    this.addSpeech("with");
     this.addSpeech(ring.getSubSystems().size());
-    this.addSpeech("subrings");
+    this.addSpeech("個の");
+    this.addSpeech("部分環"); // subrings
+    this.addSpeech("を"); 
+    this.addSpeech("含有"); // with 
+    this.addSpeech("、"); // Punctuation
     this.describeSubstitutions(ring);
   }
 
 
   @Override
   public void visit(final RichSubRing ring) {
-    this.addSpeech("Subring");
-    this.addName(ring);
+    this.addSpeech(ring.getComponentsPositions().size());
+    this.addSpeech("員"); // Elements
+    this.addSpeech("部分環"); // Subring
+    this.addSpeech("、"); // Punctuation
+    if (this.shortDescription) {
+      return;
+    }
+    this.describeReplacements(ring);
+    this.describeMultiBonds(ring);
   }
 
 
   @Override
   public void visit(final RichAliphaticChain chain) {
-    this.addName(chain);
+    // this.addSpeech("脂肪鎖"); // Aliphatic chain
+    this.addSpeech("長さ"); // length
+    this.addSpeech(chain.getComponentsPositions().size());
+    this.addSpeech("の"); // of
+    this.addSpeech("直鎖"); // Chain
+    this.addSpeech("、"); // Punctuation
     if (this.shortDescription) {
       return;
     }
+    this.describeReplacements(chain);
+    this.describeMultiBonds(chain);
     this.describeSubstitutions(chain);
   }
 
 
   @Override
   public void visit(final RichFunctionalGroup group) {
-    this.addSpeech("Functional group");
-    this.addName(group);
+    this.addSpeech("官能基");
+    this.addSpeech(group.getStructuralFormula());
+    this.addSpeech("、"); // Punctuation
   }
 
 
   @Override
   public void visit(final RichMolecule molecule) {
-    this.addName(molecule);
+    this.shortDescription = true;
+    Integer i = 0;
+    for (String set : molecule.getPath()) {
+      ((RichChemObject)
+       RichStructureHelper.getRichStructure(set)).accept(this);
+      i++;
+      if (i == 1) {
+        this.remSpeech();
+        this.addSpeech("と、"); // and Punctuation
+      }
+    }
+    this.remSpeech();
+    this.addSpeech("で構成された分子");  // Molecule consisting of
+    this.addSpeech("、"); // Punctuation
+    this.shortDescription = false;
   }
 
 
   @Override
   public void visit(final SpiroAtom spiroAtom) {
     this.shortDescription = true;
-    this.addSpeech("spiro atom");
     RichStructureHelper.getRichAtom(spiroAtom.getConnector()).accept(this);
-    this.addSpeech("to");
+    this.addSpeech("スピロ原子"); // spiro atom
+    this.addSpeech("に"); // to 
     RichStructureHelper.getRichAtomSet(spiroAtom.getConnected()).accept(this);
+    this.addSpeech("、"); // Punctuation
     this.shortDescription = false;
   }
 
 
   @Override
   public void visit(final BridgeAtom bridgeAtom) {
-    this.addSpeech("bridge atom");
     RichStructureHelper.getRichAtom(bridgeAtom.getConnector()).accept(this);
+    this.addSpeech("橋頭原子");  // bridge atom
   }
 
 
   @Override
   public void visit(final ConnectingBond bond) {
     this.shortDescription = true;
-    RichStructureHelper.getRichBond(bond.getConnector()).accept(this);
-    // TODO (sorge) The past tense here is problematic!
-    this.modSpeech("ed");
-    this.addSpeech("to");
+    this.subject = false;
     String connected = bond.getConnected();
     if (RichStructureHelper.isAtom(connected)) {
       RichStructureHelper.getRichAtom(connected).accept(this);
     } else {
       RichStructureHelper.getRichAtomSet(connected).accept(this);
     }
+    this.addSpeech("に"); // to 
+    RichStructureHelper.getRichBond(bond.getConnector()).accept(this);
+    this.addSpeech("、"); // Punctuation
+    // TODO (sorge) The past tense here is problematic!
+    // this.modSpeech("して"); // ed (modifier)
     this.shortDescription = false;
+    this.subject = true;
   }
 
 
   @Override
   public void visit(final SharedAtom sharedAtom) {
     this.shortDescription = true;
-    this.addSpeech("shared atom");
     RichStructureHelper.getRichAtom(sharedAtom.getConnector()).accept(this);
-    this.addSpeech("with");
+    this.addSpeech("共有原子"); // shared atom
     RichStructureHelper.getRichAtomSet(sharedAtom.getConnected()).accept(this);
+    this.addSpeech("含有"); // with 
+    this.addSpeech("、"); // Punctuation
     this.shortDescription = false;
   }
 
 
   @Override
   public void visit(final SharedBond sharedBond) {
-    this.addSpeech("shared");
     RichStructureHelper.getRichBond(sharedBond.getConnector()).accept(this);
+    this.addSpeech("共有"); // shared
   }
 
 
@@ -199,23 +260,61 @@ public class EnglishExpertSpeechVisitor extends AbstractSpeechVisitor {
   }
 
 
-  // TODO (sorge) Do something about all upper case names without destroying
-  // important upper cases. E.g.: WordUtils.capitalizeFully.
-  private void addName(final RichAtomSet atomset) {
-    if (!atomset.getName().equals("")) {
-      addSpeech(atomset.getName());
-      return;
-    }
-    if (!atomset.getIupac().equals("")) {
-      addSpeech(atomset.getIupac());
-      return;
-    }
-    addSpeech(atomset.getMolecularFormula());
+  public String getSpeech() {
+    final Joiner joiner = Joiner.on("");
+    String result = joiner.join(this.retrieveSpeech());
+    this.clearSpeech();
+    return result;
   }
 
 
   // TODO (sorge) For the following utility functions, see if they can be
   // refactored with walk methods, etc.
+  private void describeReplacements(final RichAtomSet system) {
+    final Iterator<String> iterator = system.iterator();
+    while (iterator.hasNext()) {
+      final String value = iterator.next();
+      final RichAtom atom = RichStructureHelper.getRichAtom(value);
+      if (!atom.isCarbon()) {
+        this.addSpeech(system.getPosition(value));
+        this.addSpeech("位"); // Position symbol
+        this.addSpeech("は"); // at 
+        this.addSpeech(atom.getName());
+        this.addSpeech("、"); // Punctuation
+      }
+    }
+  }
+
+
+  //TODO (sorge) Sort those bonds. Maybe combine with a more stateful walk.
+  private void describeMultiBonds(final RichAtomSet system) {
+    for (final String component : system.getComponents()) {
+      if (!RichStructureHelper.isBond(component)) {
+        continue;
+      }
+      RichBond bond = RichStructureHelper.getRichBond(component);
+      if (bond.isSingle()) {
+        continue;
+      }
+      Integer atomA = system.getPosition(bond.getComponents().first());
+      Integer atomB = system.getPosition(bond.getComponents().last());
+      if (atomA > atomB) {
+        atomA ^= atomB;
+        atomB ^= atomA;
+        atomA ^= atomB;
+      }
+      this.addSpeech(atomA);
+      this.addSpeech("位"); // Position symbol
+      this.addSpeech("と"); // and
+      this.addSpeech(atomB);
+      this.addSpeech("位"); // Position symbol
+      this.addSpeech("の間は"); // between
+      bond.accept(this);
+      this.addSpeech("、"); // Punctuation
+    }
+  }
+
+
   private void describeSubstitutions(final RichAtomSet system) {
     final SortedSet<Integer> subst = new TreeSet<Integer>();
     for (final String atom : system.getConnectingAtoms()) {
@@ -225,39 +324,38 @@ public class EnglishExpertSpeechVisitor extends AbstractSpeechVisitor {
       case 0:
         return;
       case 1:
-        this.addSpeech("Substitution at position");
-        this.addSpeech(subst.iterator().next());
-        return;
       default:
-        this.addSpeech("Substitutions at positions");
         for (final Integer position : subst) {
           this.addSpeech(position);
-          this.addSpeech("and");
+          this.addSpeech("位"); // position
+          this.addSpeech("と"); // and
         }
         this.remSpeech();
+        this.addSpeech("で"); // at
+        this.addSpeech("置換"); // Substitution
+        this.addSpeech("、"); // Punctuation
+        return;
     }
   }
 
-  /**
+
+  /** 
    * Adds description of hydrogen bonds of an atom.
-   *
+   * 
    * @param atom
    *          The atom to describe.
    */
-  private void describeHydrogenBonds(final RichAtom atom) {
+  private void describeHydrogenBonds(RichAtom atom) {
     final Integer count = atom.getStructure().getImplicitHydrogenCount();
     switch (count) {
       case 0:
         return;
       case 1:
-        this.addSpeech("bonded to");
-        this.addSpeech(count.toString());
-        this.addSpeech("hydrogen");
-        return;
       default:
-        this.addSpeech("bonded to");
+        this.addSpeech("水素");  // hydrogen (and hydrogens)
         this.addSpeech(count.toString());
-        this.addSpeech("hydrogens");
+        // this.addSpeech("に結合しており、"); // bonded to
+        this.addSpeech("に結合、"); // bonded to
         return;
     }
   }
