+++ src/main/java/com/progressiveaccess/cmlspeech/speech/Language.java
@@ -25,7 +25,7 @@
 
 //
 
-package com.progressiveaccess.cmlspeech.sre;
+package com.progressiveaccess.cmlspeech.speech;
 
 import com.progressiveaccess.cmlspeech.base.Cli;
